+++ Assets/_Project/Scripts/Gameplay/Projectile/RSE_OnDamakeTaken.cs
@@ -4,9 +4,9 @@ using UnityEngine.Events;
 [CreateAssetMenu(fileName = "RSE_OnDamakeTaken", menuName = "Scriptable Objects/RSE_OnDamakeTaken")]
 public class RSE_OnDamakeTaken : ScriptableObject
 {
-    public UnityAction<int> OnEventRaised;
+    public UnityAction<float> OnEventRaised;
 
-    public void RaiseEvent(int amount)
+    public void RaiseEvent(float amount)
     {
         if (OnEventRaised != null)
         {
